lemonldap CVEs
This one is really weird because the various Perl
distributions are embedded in lemonldap instead 
of the other way around. 

Also, the lemonldap stuff is on CPAN, but perhaps
incomplete or dated. This means that we will need
to do something extra to map its modules to distros
for the database.

This will do until we can get something better.