Add metadata to libtommath for CryptX
Note that these versions are approximate, because the documentation does
not often indicate what commit or even what date the library has been
updated to.

libtomcrypt has also had a long commit history in the development branch
before versions were issued, according to the changelog in their repo
(which was omitted from CryptX, making it much harder to determine
dependencies).